fix: 네이밍 Detail -> GuestBook으로 수정
