--- android/feature/invitation/src/main/kotlin/com/andlife/invitation/model/detail/InvitationDetailSideEffect.kt
@@ -1,9 +1,0 @@
-package com.andlife.invitation.model.detail
-
-import com.andlife.ui.base.BaseSideEffect
-
-sealed interface InvitationDetailSideEffect : BaseSideEffect {
-    data class ShowSnackbar(
-        val message: String,
-    ) : InvitationDetailSideEffect
-}
