--- android/feature/invitation/src/main/kotlin/com/andlife/invitation/model/detail/InvitationDetailUiEvent.kt
@@ -1,22 +1,0 @@
-package com.andlife.invitation.model.detail
-
-import com.andlife.ui.base.BaseUiEvent
-import com.andlife.ui.component.invitation.SelectedMedia
-
-sealed interface InvitationDetailUiEvent : BaseUiEvent {
-    data class UpdateSelectedMedias(
-        val medias: List<SelectedMedia>,
-    ) : InvitationDetailUiEvent
-
-    data class UpdateTextContent(
-        val textContent: String,
-    ) : InvitationDetailUiEvent
-
-    data class RemoveMedia(
-        val media: SelectedMedia,
-    ) : InvitationDetailUiEvent
-
-    data object UploadMedias : InvitationDetailUiEvent
-
-    data object ClearError : InvitationDetailUiEvent
-}
