+++ android/feature/invitation/src/main/kotlin/com/andlife/invitation/model/guestbook/InvitationGuestBookSideEffect.kt
@@ -1,0 +1,9 @@
+package com.andlife.invitation.model.guestbook
+
+import com.andlife.ui.base.BaseSideEffect
+
+sealed interface InvitationGuestBookSideEffect : BaseSideEffect {
+    data class ShowSnackbar(
+        val message: String,
+    ) : InvitationGuestBookSideEffect
+}
